requirements.txt
# /Users/thanhhuyen/Documents/deplot/predict-viral-content/requirements.txt
praw
azure-eventhub
python-dotenv
